extract sharable relationships job code to parent class
