+++ app/jobs/bulkrax/child_relationships_job.rb
@@ -2,9 +2,7 @@
 
 module Bulkrax
   class ChildWorksError < RuntimeError; end
-  class ChildRelationshipsJob < ApplicationJob
-    queue_as :import
-
+  class ChildRelationshipsJob < RelationshipsJob
     def perform(*args)
       @args = args
 
@@ -42,10 +40,6 @@ module Bulkrax
       work_parent_work_child(members_works) if members_works.present?
     end
 
-    def entry
-      @entry ||= Bulkrax::Entry.find(@args[0])
-    end
-
     def child_entries
       @child_entries ||= @args[1].map { |e| Bulkrax::Entry.find(e) }
     end
@@ -59,70 +53,7 @@ module Bulkrax
       end
     end
 
-    def importer_run_id
-      @args[2]
-    end
-
-    def user
-      @user ||= entry.importerexporter.user
-    end
-
     private
-
-    # rubocop:disable Rails/SkipsModelValidations
-    # Work-Collection membership is added to the child as member_of_collection_ids
-    # This is adding the reverse relatinship, from the child to the parent
-    def work_child_collection_parent(work_id)
-      attrs = { id: work_id, collections: [{ id: entry&.factory&.find&.id }] }
-      Bulkrax::ObjectFactory.new(attributes: attrs,
-                                 source_identifier_value: child_works_hash[work_id][entry.parser.source_identifier],
-                                 work_identifier: entry.parser.work_identifier,
-                                 collection_field_mapping: entry.parser.collection_field_mapping,
-                                 replace_files: false,
-                                 user: user,
-                                 klass: child_works_hash[work_id][:class_name].constantize).run
-      ImporterRun.find(importer_run_id).increment!(:processed_children)
-    rescue StandardError => e
-      entry.status_info(e)
-      ImporterRun.find(importer_run_id).increment!(:failed_children)
-    end
-
-    # Collection-Collection membership is added to the as member_ids
-    def collection_parent_collection_child(member_ids)
-      attrs = { id: entry&.factory&.find&.id, children: member_ids }
-      Bulkrax::ObjectFactory.new(attributes: attrs,
-                                 source_identifier_value: entry.identifier,
-                                 work_identifier: entry.parser.work_identifier,
-                                 collection_field_mapping: entry.parser.collection_field_mapping,
-                                 replace_files: false,
-                                 user: user,
-                                 klass: entry.factory_class).run
-      ImporterRun.find(importer_run_id).increment!(:processed_children)
-    rescue StandardError => e
-      entry.status_info(e)
-      ImporterRun.find(importer_run_id).increment!(:failed_children)
-    end
-
-    # Work-Work membership is added to the parent as member_ids
-    def work_parent_work_child(member_ids)
-      # build work_members_attributes
-      attrs = { id: entry&.factory&.find&.id,
-                work_members_attributes: member_ids.each.with_index.each_with_object({}) do |(member, index), ids|
-                  ids[index] = { id: member }
-                end }
-      Bulkrax::ObjectFactory.new(attributes: attrs,
-                                 source_identifier_value: entry.identifier,
-                                 work_identifier: entry.parser.work_identifier,
-                                 collection_field_mapping: entry.parser.collection_field_mapping,
-                                 replace_files: false,
-                                 user: user,
-                                 klass: entry.factory_class).run
-      ImporterRun.find(importer_run_id).increment!(:processed_children)
-    rescue StandardError => e
-      entry.status_info(e)
-      ImporterRun.find(importer_run_id).increment!(:failed_children)
-    end
-    # rubocop:enable Rails/SkipsModelValidations
 
     def reschedule(entry_id, child_entry_ids, importer_run_id)
       ChildRelationshipsJob.set(wait: 10.minutes).perform_later(entry_id, child_entry_ids, importer_run_id)
